+++ pytsite/comments/_error.py
@@ -1,0 +1,13 @@
+"""PytSite Comments Package Errors.
+"""
+__author__ = 'Alexander Shepetko'
+__email__ = 'a@shepetko.com'
+__license__ = 'MIT'
+
+
+class DriverAlreadyRegistered(Exception):
+    pass
+
+
+class DriverNotRegistered(Exception):
+    pass