+++ pytsite/update/_command.py
@@ -42,7 +42,7 @@ class Update(_console.command.Abstract):
             for minor in range(0, 100):
                 if stop:
                     break
-                for rev in range(0, 100):
+                for rev in range(0, 20):
                     if stop:
                         break